requirements.txt
fastapi
uvicorn
spectral
scikit-learn
tensorflow
absl-py
joblib
python-multipart
